requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
sqlalchemy-cockroachdb>=2.0.0